Require users to login for views and allow them to logout.
Requiring users to login means redirecting users to the login view when
attempting to access a page that requires logging in. Redirecting users
to the login view interrupts the user's workflow which means we should
be nice and redirect them back to the page they were trying to access
when they were interrupted. In order to do this we need to handle the
next argument provided to the request query parameters in the login
route.

This commit also provides user's a way to logout. The logout button is
context dependent on the current_user's logged in state.